Split app.js phase 4c-1: extract enforcement module (~3.5k lines)
Migration/onboarding banner state (12 vars) moved to state.js. In-app
tests pass (132/132).
